Code refactoring in the entrepreneur app views
