+++ app/static/sass/entrepreneur/general_venture_form.scss
