+++ app/templates/layout/base_email.html
@@ -26,6 +26,10 @@
             </td>
         </tr>
 
+        <tr>
+            <a href="{{ base_url }}account/update/" style="color:gray;">Go to manage email notifications.</a>
+        </tr>
+
         <tr style="text-align: center;">
             <td style="padding: 20px 0px 5px">
                 <a href="https://www.facebook.com/QueryJane-1788853644723154/" target="_blank"><img src="https://s3-us-west-2.amazonaws.com/queryjane/social/facebook.png" width="16px" height="16px" style="padding: 0px 9px" /></a>
