+++ entrepreneur/templates/entrepreneur/venture_settings/_membership_line.html
@@ -1,8 +1,8 @@
 {% load static %}
 <li class="media">
     <div class="media-left">
-        <a href="#">
-            <img class="media-object" src="http://placehold.it/50x50" alt="...">
+        <a href="{% url 'professional_detail' membership.admin.slug %}">
+            <img style="width: 50px; height: 50px" class="media-object thumbnail" src="{{ membership.admin.user.get_avatar }}" alt="{{ membership.admin }}">
         </a>
     </div>
