+++ entrepreneur/templates/entrepreneur/venture_settings/general_venture_form.html
@@ -4,7 +4,7 @@
 {% block title %}{{ venture }} - Settings{% endblock %}
 
 {% block css %}
-<link href="{% static 'dist/css/entrepreneur/settings_venture_form.css' %}" rel="stylesheet">
+<link href="{% static 'dist/css/entrepreneur/general_venture_form.css' %}" rel="stylesheet">
 {% endblock %}
 
 {% block content %}
@@ -151,5 +151,5 @@
 {% endblock %}
 
 {% block js %}
-<script src="{% static 'js/entrepreneur/settings_venture_form.js' %}"></script>
+<script src="{% static 'js/entrepreneur/general_venture_form.js' %}"></script>
 {% endblock %}
