+++ entrepreneur/templates/entrepreneur/venture_settings/job_form.html
@@ -68,7 +68,7 @@
                             <br>
 
                             <p class="h4">Where is the job?</p>
-                            <p>Fill these fields only if it is necessary for the apsirantes to live or move to a specific country or city.</p>
+                            <p>Fill these fields only if it is necessary for the applicant to live or move to a specific country or city.</p>
 
                             <div class="row">
                                 <div class="col-md-6 col-xs-12">
