+++ entrepreneur/templates/entrepreneur/venture_settings/settings_menu.html
@@ -10,7 +10,7 @@
 
     <ul class="list-group QjaneVentureSettingsMenu">
         <li class="list-group-item {% if general_active %}active{% endif %}">
-            <a href="{% url 'entrepreneur:settings_venture_form' venture.slug %}" class="clearfix">
+            <a href="{% url 'entrepreneur:general_venture_form' venture.slug %}" class="clearfix">
                 <span class="pull-left">General</span>
                 <i class="fa fa-cog pull-right" aria-hidden="true"></i>
             </a>
